Split `ParallelBlockMerge.h` into several headers
Pure refactoring, no functional change. The single 712-line header is
split into `MergeHelpers.h` (`computeSplitters` and the `detail`
helpers), `ChunkMerger.h`, `SerialMergeState.h`, and
`ParallelMergeState.h`. `ParallelBlockMerge.h` keeps only the public
`parallelBlockMergeToRange` and includes the new headers, so that it
remains the single header that users of the library include. All code is
moved verbatim, and each new header is self-contained.
